Fix bug: task-complete fails to load filter
Bug: While a filter is being loaded, after completing a task,the panel
would fail to load the updated filtered tasks. It is the handler for
complete button only supported loading user-projects but not filters.

Solution: Added validation process into the complete button handler,
where it differentiates the type of current load project title. It
calls `loadTaskFilter()` if it's a filter, calls `loadCurrentProject()`
if it's a user-project.

Additional improvement:
- Renamed `loadTaskFilters` to `taskFilterLoaders`.